+++ protocols/directfund/directfund.go
@@ -447,7 +447,7 @@ func IsDirectFundObjective(id protocols.ObjectiveId) bool {
 // There is no AppDefinition, since we currently only support full consensus rules for direct channels.
 type ObjectiveRequestForConsensusApp struct {
 	CounterParty      types.Address
-	ChallengeDuration uint64
+	ChallengeDuration uint32
 	Outcome           outcome.Exit
 	Nonce             uint64
 }
